CSL-245 Add PDF conversion to submit-request
Refactor previous PDF processing
Import and use prepareUpload from Notify
Add generated PDF to user email
Remove now unused submit behaviour
